+++ auth/data.js
@@ -1,206 +1,148 @@
-var debug   = require('debug')('auth:data:client');
+var debug   = require('debug')('auth:data');
 var base    = require('taskcluster-base');
 var assert  = require('assert');
 var _       = require('lodash');
 
-/** Configure a client Entity subclass */
-var Client = base.LegacyEntity.configure({
-  mapping: [
-    {
-      key:              'PartitionKey',
-      property:         'clientId',
-      type:             'string'
-    }, {
-      // This is always hardcoded to 'credentials'
-      key:              'RowKey',
-      type:             'string',
-      hidden:           true
-    }, {
-      key:              'accessToken',
-      type:             'string'
-    }, {
-      key:              'name',
-      type:             'string'
-    }, {
-      key:              'version',
-      type:             'string'
-    }, {
-      key:              'scopes',
-      type:             'json'
-    }, {
-      key:              'expires',
-      type:             'date'
-    }, {
-      key:              'details',
-      type:             'json'
-    }
-  ]
+var Client = base.Entity.configure({
+  version:          1,
+  partitionKey:     base.Entity.keys.StringKey('clientId'),
+  rowKey:           base.Entity.keys.ConstantKey('client'),
+  signEntities:     true,
+  properties: {
+    clientId:       base.Entity.types.String,
+    description:    base.Entity.types.Text,
+    accessToken:    base.Entity.types.EncryptedText,
+    expires:        base.Entuty.types.Date,
+    /**
+     * Details object with properties:
+     * - created          // Time when client was created
+     * - lastModified     // Last time client was modified
+     * - lastDateUsed     // Only updated if more than 6 hours out of date
+     * - lastRotated      // Last time accessToken was reset
+     * (more properties may be added in the future)
+     */
+    details:        base.Entity.types.JSON
+  }
 });
 
-// RowKey constant, used as we don't need a RowKey
-var ROW_KEY_CONST = 'credentials';
-
-/** Create a client */
-Client.create = function(properties) {
-  properties.RowKey = ROW_KEY_CONST;
-  return base.LegacyEntity.create.call(this, properties);
-};
-
-/** Load client from clientId */
-Client.load = function(clientId) {
-  return base.LegacyEntity.load.call(this, clientId, ROW_KEY_CONST);
-};
-
-/** Load all clients */
-Client.loadAll = function() {
-  return base.LegacyEntity.queryRowKey.call(this, ROW_KEY_CONST);
-};
-
-/** Remove client with given clientId */
-Client.remove = function(clientId) {
-  return base.LegacyEntity.remove.call(this, clientId, ROW_KEY_CONST);
-};
-
-/** Create a clientLoader that works with base.API.createSignatureValidator */
-Client.createClientLoader = function() {
-  var Client = this;
-  return async (clientId) => {
-    try {
-      var client = await Client.load(clientId);
-    } catch (err) {
-      debug('Failed to load client: ' + clientId);
-      throw new Error("Failed to load client");
-    }
-    if (client.expires.getTime() < new Date().getTime()) {
-      throw new Error("Credentials expired!");
-    }
-    return {
-      clientId:     client.clientId,
-      accessToken:  client.accessToken,
-      scopes:       client.scopes
-    };
-  };
-};
-
-/** Create caching client loader */
-Client.createCachedClientLoader = function(options) {
-  options = _.defaults(options || {}, {
-    cacheTimeout:       10 * 60 * 60 * 1000
-  });
-  assert(typeof(options.cacheTimeout) == 'number',
-         "Expected options.cacheTimeout to be a number!");
-
-  var clientLoader = this.createClientLoader();
-  var cache = {};
-  setInterval(() => {
-    // clean up cache
-    var now = new Date().getTime();
-    _.keys(cache).forEach(clientId => {
-      if (cache[clientId].reloadAt < now) {
-        delete cache[clientId];
-      }
-    });
-  }, options.cacheTimeout);
-
-  return async (clientId) => {
-    var now = new Date().getTime();
-    var entry = cache[clientId];
-    if (!entry || entry.reloadAt < now) {
-      var error = null, client = null;
-      try {
-        client = await clientLoader(clientId);
-      } catch (err) {
-        error = err;
-      }
-      cache[clientId] = entry = {
-        client,
-        error,
-        reloadAt: now + options.cacheTimeout
-      };
-    }
-    if (entry.error) {
-      throw entry.error;
-    }
-    return entry.client;
+/** Get JSON representation of client */
+Client.prototype.json = function() {
+  return {
+    clientId:       this.clientId,
+    description:    this.description,
+    expires:        this.expires.toJSON(),
+    created:        this.details.created,
+    lastModified:   this.details.lastModified,
+    lastDateUsed:   this.details.lastDateUsed,
+    lastRotated:    this.details.lastRotated
   };
 };
 
 /**
- * Create client with root credentials
+ * Ensure root client exists and has the given accessToken.
+ * If accessToken has changed (or isn't set) root client will be overwritten,
+ * and expires set 30 min into the future.
  *
- * The root client is hardcoded into the configuration and is used to create
- * other users. Technically, it's possible to modify the root user, but all
- * changes will be forgotten as we reset whenever the server starts.
+ * Should only be called if the app is configured with a rootAccessToken.
+ * Otherwise, app should assume whatever is in the table storage is the
+ * root access token, and that appropriate role is attached.
  *
- * credentials: {
- *   clientId:        '...',
- *   accessToken:     '...'
- * }
+ * Basically, this is for bootstrapping only.
  */
-Client.createRootClient = function(credentials) {
-  assert(credentials.clientId,    "root credentials must have clientId");
-  assert(credentials.accessToken, "root credentials must have accessToken");
-  var values = {
-    version:      '0.2.0',
-    clientId:     credentials.clientId,
-    accessToken:  credentials.accessToken,
-    scopes:       ['*'],
-    expires:      new Date(3000, 0, 1), // 1st of January year 3000
-    name:         "Root Client",
-    details: {
-      notes: [
-        "This is the initial user that is **hardcoded** into the configuration",
-        "of auth.taskcluster.net. Note, that these credentials should only be",
-        "used to create the other taskcluster credentials.",
-        "",
-        "**Do not modify** this user, all changes will be reset whenever the",
-        "server process restarts (this is often)."
-      ].join('\n')
-    }
-  };
-  // Load root client (if it exists)
-  var Client = this;
-  return Client.load(credentials.clientId).then(function(client) {
-    // If it exists we make sure to modify it (reset any changes)
-    return client.modify(function() {
-      this.accessToken  = values.accessToken;
-      this.scopes       = values.scopes;
-      this.expires      = values.expires;
-      this.name         = values.name;
-      this.details      = values.details;
+Client.ensureRootClient = async function(accessToken) {
+  assert(typeof(accessToken) === 'string',
+         "Expected accessToken to be a string");
+  let Client = this;
+
+  let client = await Client.load({clientId: 'root'}, true);
+  if (client) {
+    return client.modify(client => {
+      if (client.accessToken !== accessToken) {
+        client.accessToken            = accessToken;
+        client.description            = "Automatically created `root` client " +
+                                        "for bootstrapping API access.";
+        client.expires                = taskcluster.fromNow('30 min'),
+        client.details.created        = new Date().toJSON();
+        client.details.lastModified   = new Date().toJSON();
+        client.details.lastDateUsed   = new Date().toJSON();
+        client.details.lastRotated    = new Date().toJSON();
+      }
     });
-  }, function(err) {
-    // if it doesn't exist then we'll try to create it
-    if (err.code === 'ResourceNotFound') {
-      return Client.create(values).then(undefined, function(err) {
-        // If it already exists, then we've been racing with another process
-        // to create it, so we'll just modify it, I realize that we've been
-        // doing this... but it makes sense to always assume it exist initially
-        // as this will be the typical case. And then it makes sense to ensure
-        // against the case were we're racing...
-        if (err.code === 'EntityAlreadyExists') {
-          return Client.load(credentials.clientId).then(function(client) {
-            // If it exists we make sure to modify it (reset any changes)
-            return client.modify(function() {
-              this.accessToken  = values.accessToken;
-              this.scopes       = values.scopes;
-              this.expires      = values.expires;
-              this.name         = values.name;
-              this.details      = values.details;
-            });
-          });
-        }
-        throw err;
-      });
+  }
+
+  // Create client resolving conflicts by overwriting
+  await Client.create({
+    clientId:         'root',
+    description:      "Automatically created `root` client " +
+                      "for bootstrapping API access";
+    accessToken:      accessToken,
+    expires:          taskcluster.fromNow('30 min'),
+    details: {
+      created:        new Date().toJSON(),
+      lastModified:   new Date().toJSON(),
+      lastDateUsed:   new Date().toJSON(),
+      lastRotated:    new Date().toJSON()
     }
-    throw err;
-  }).then(function() {
-    debug("Created root credentials");
-  }, function(err) {
-    debug("Failed to create root credentials, err: %s, json: %j", err, err);
-    throw err;
-  });
+  }, true);
 };
 
-// Export client
+// Export Client
 exports.Client = Client;
+
+var Role = base.Entity.configure({
+  version:          1,
+  partitionKey:     base.Entity.keys.StringKey('roleId'),
+  rowKey:           base.Entity.keys.ConstantKey('role'),
+  signEntities:     true,
+  properties: {
+    roleId:         base.Entity.types.String,
+    description:    base.Entity.types.Text,
+    scopes:         base.Entity.types.JSON,
+    /**
+     * Details object with properties:
+     * - created
+     * - lastModified
+     * (more properties may be added in the future)
+     */
+    details:        base.Entity.types.JSON,
+  }
+});
+
+/** Get JSON representation of client */
+Role.prototype.json = function(resolver) {
+  return {
+    roleId:         this.roleId,
+    description:    this.description,
+    created:        this.details.created,
+    lastModified:   this.details.lastModified,
+    scopes:         this.scopes,
+    expandedScopes: resolver.resolve(this.scopes)
+  };
+};
+
+/**
+ * Ensure the role: client-id:root -> ['*'] exists
+ *
+ * Should only be called if the app is configured with a rootAccessToken.
+ * Otherwise, app should assume whatever is in the table storage is the
+ * root access token, and that appropriate role is attached.
+ *
+ * Basically, this is for bootstrapping only.
+ */
+Role.ensureRootRole = function() {
+  let Role = this;
+  return Role.create({
+    roleId:       'client-id:root',
+    description:  "Automatically created role for bootstrapping the `root` "+
+                  "client.",
+    scopes:       ['*'],
+    details: {
+      created:        new Date().toJSON(),
+      lastModified:   new Date().toJSON()
+    }
+  }, true);
+};
+
+// Export Role
+exports.Role = Role;
+
