+++ bin/server.js
@@ -1,29 +1,35 @@
 #!/usr/bin/env node
-var base    = require('taskcluster-base');
-var data    = require('../auth/data');
-var v1      = require('../routes/api/v1');
-var path    = require('path');
-var debug   = require('debug')('taskcluster-auth:bin:server');
-var Promise = require('promise');
-var persona = require('../auth/persona');
-var AWS     = require('aws-sdk-promise');
+var base          = require('taskcluster-base');
+var data          = require('../auth/data');
+var v1            = require('../auth/v1');
+var path          = require('path');
+var debug         = require('debug')('server');
+var Promise       = require('promise');
+var persona       = require('../auth/persona');
+var AWS           = require('aws-sdk-promise');
+var exchanges     = require('../auth/exchanges');
+var ScopeResolver = require('../auth/scoperesolver');
+var taskcluster   = require('taskcluster-client');
 
 /** Launch server */
-var launch = function(profile) {
+var launch = async function(profile) {
   // Load configuration
   var cfg = base.config({
     defaults:     require('../config/defaults'),
     profile:      require('../config/' + profile),
     envs: [
+      'auth_tableSigningKey',
+      'auth_tableCryptoKey',
       'server_publicUrl',
       'server_cookieSecret',
       'azure_accountName',
       'azure_accountKey',
+      'pulse_username',
+      'pulse_password',
       'aws_accessKeyId',
       'aws_secretAccessKey',
       'influx_connectionString',
-      'auth_root_clientId',
-      'auth_root_accessToken',
+      'auth_rootAccessToken',
       'auth_azureAccounts',
       'auth_clientIdForTempCreds'
     ],
@@ -47,86 +53,126 @@ var launch = function(profile) {
   // Configure client table
   var Client = data.Client.configure({
     tableName:    cfg.get('auth:clientTableName'),
-    credentials:  cfg.get('azure')
+    credentials:  cfg.get('azure'),
+    signingKey:   cfg.get('auth:tableSigningKey'),
+    cryptoKey:    cfg.get('auth:tableCryptoKey')
   });
 
-  // Initialize validator and publish schemas if needed
-  var validator = null;
-  var validatorLoaded = base.validator({
-    folder:           path.join(__dirname, '..', 'schemas'),
-    constants:        require('../schemas/constants'),
-    publish:          cfg.get('auth:publishMetaData') === 'true',
-    schemaPrefix:     'auth/v1/',
-    aws:              cfg.get('aws')
-  }).then(function(validator_) {
-    validator = validator_;
+  // Configure role table
+  var Role = data.Role.configure({
+    tableName:    cfg.get('auth:rolesTableName'),
+    credentials:  cfg.get('azure'),
+    signingKey:   cfg.get('auth:tableSigningKey'),
+    cryptoKey:    cfg.get('auth:tableCryptoKey')
   });
 
-  // Load validator and create client table before proceeding
-  return Promise.all([
-    validatorLoaded,
-    Client.createTable().then(function() {
-      // Create root credentials
-      return Client.createRootClient(cfg.get('auth:root'));
+  var validator, publisher, exchangePrefix;
+  await Promise.all([
+    (async () => {
+      // Initialize validator and publish schemas if needed
+      validator = await base.validator({
+        folder:           path.join(__dirname, '..', 'schemas'),
+        constants:        require('../schemas/constants'),
+        publish:          cfg.get('auth:publishMetaData') === 'true',
+        schemaPrefix:     'auth/v1/',
+        aws:              cfg.get('aws')
+      });
+
+      publisher = await exchanges.setup({
+        credentials:      cfg.get('pulse'),
+        exchangePrefix:   cfg.get('auth:exchangePrefix'),
+        validator:        validator,
+        referencePrefix:  'auth/v1/exchanges.json',
+        publish:          cfg.get('auth:publishMetaData') === 'true',
+        aws:              cfg.get('aws'),
+        drain:            influx,
+        component:        cfg.get('auth:statsComponent'),
+        process:          'server'
+      });
+
+      exchangePrefix = exchanges.reference({
+        credentials:      cfg.get('pulse'),
+        exchangePrefix:   cfg.get('auth:exchangePrefix')
+      }).exchangePrefix;
+    })(),
+    (async() {
+      // Ensure tables exist
+      await Promise.all([
+        Client.ensureTable(),
+        Role.ensureTable()
+      ]);
+
+      if (cfg.get('auth:rootAccessToken')) {
+        await Client.ensureRootClient(cfg.get('auth:rootAccessToken'));
+        await Role.ensureRootRole();
+      }
     })
-  ]).then(function() {
-    var signatureValidator = base.API.createSignatureValidator({
-      clientLoader: Client.createCachedClientLoader({
-        cacheTimeout:       cfg.get('auth:clientCacheTimeout')
-      })
-    });
+  ]);
 
-    // Create API router and publish reference if needed
-    return v1.setup({
-      context: {
-        validator,
-        Client,
-        signatureValidator,
-        azureAccounts:      JSON.parse(cfg.get('auth:azureAccounts')),
-        sts:                new AWS.STS(cfg.get('aws')),
-        rootClientId:       cfg.get('auth:root:clientId'),
-      },
-      validator:          validator,
-      signatureValidator: signatureValidator,
-      publish:            cfg.get('auth:publishMetaData') === 'true',
-      baseUrl:            cfg.get('server:publicUrl') + '/v1',
-      referencePrefix:    'auth/v1/api.json',
-      aws:                cfg.get('aws'),
-      component:          cfg.get('auth:statsComponent'),
-      drain:              influx
-    });
-  }).then(function(router) {
-    // Create app
-    var app = base.app({
-      port:           Number(process.env.PORT || cfg.get('server:port')),
-      env:            cfg.get('server:env'),
-      forceSSL:       cfg.get('server:forceSSL'),
-      trustProxy:     cfg.get('server:trustProxy')
-    });
-
-    // Mount API router
-    app.use('/v1', router);
-
-
-    // Setup middleware and authentication
-    var ensureAuth = persona.setup(app, {
-      cookieSecret:   cfg.get('server:cookieSecret'),
-      viewFolder:     path.join(__dirname, '..', 'views'),
-      assetFolder:    path.join(__dirname, '..', 'assets'),
-      development:    cfg.get('server:development') === 'true',
-      publicUrl:      cfg.get('server:publicUrl')
-    });
-
-    // Provide a client
-    app.globals = {
-      Client:               Client,
-      root:                 cfg.get('auth:root'),
-      clientIdForTempCreds: cfg.get('auth:clientIdForTempCreds')
-    };
-
-    // Create server
-    return app.createServer();
+  // Create resolver
+  var resolver = new ScopeResolver({
+    Client,
+    Role,
+    exchangePrefix,
+    connection: new taskcluster.PulseConnection({
+      credentials: cfg.get('pulse')
+    })
   });
+
+  // Load everything for resolver
+  await resolver.setup();
+
+  // Create signature validator
+  var signatureValidator = resolver.createSignatureValidator();
+
+  // Create API router and publish reference if needed
+  var router = await v1.setup({
+    context: {
+      Client, Role,
+      publisher,
+      resolver,
+      sts:                new AWS.STS(cfg.get('aws')),
+      azureAccounts:      JSON.parse(cfg.get('auth:azureAccounts')),
+      signatureValidator,
+    },
+    validator:          validator,
+    signatureValidator: signatureValidator,
+    publish:            cfg.get('auth:publishMetaData') === 'true',
+    baseUrl:            cfg.get('server:publicUrl') + '/v1',
+    referencePrefix:    'auth/v1/api.json',
+    aws:                cfg.get('aws'),
+    component:          cfg.get('auth:statsComponent'),
+    drain:              influx
+  });
+
+  // Create app
+  var app = base.app({
+    port:           Number(process.env.PORT || cfg.get('server:port')),
+    env:            cfg.get('server:env'),
+    forceSSL:       cfg.get('server:forceSSL'),
+    trustProxy:     cfg.get('server:trustProxy')
+  });
+
+  // Mount API router
+  app.use('/v1', router);
+
+  // Setup middleware and authentication
+  var ensureAuth = persona.setup(app, {
+    cookieSecret:   cfg.get('server:cookieSecret'),
+    viewFolder:     path.join(__dirname, '..', 'views'),
+    assetFolder:    path.join(__dirname, '..', 'assets'),
+    development:    cfg.get('server:development') === 'true',
+    publicUrl:      cfg.get('server:publicUrl')
+  });
+
+  // Provide a client
+  app.globals = {
+    Client:               Client,
+    clientIdForTempCreds: cfg.get('auth:clientIdForTempCreds')
+  };
+
+  // Create server
+  return app.createServer();
 };
 
 // If server.js is executed start the server
