+++ package.json
@@ -26,7 +26,7 @@
     "cookie-parser":                    "^1.3.5",
     "cookie-session":                   "^1.1.0",
     "body-parser":                      "^1.12.4",
-    "taskcluster-client":               "^0.22.1",
+    "taskcluster-client":               "^0.22.6",
     "express":                          "^4.12.4",
     "babel":                            "^5.0.9",
     "kexec":                            "^1.2.0",
